requirements.txt
psycopg2-binary>=2.9
pandas>=2.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
python-dotenv>=1.0
playwright>=1.40
